requirements-airflow.txt
apache-airflow[evidently]==3.1.0
evidently==0.7.10
lightgbm==4.6.0
joblib==1.5.1
pandas==2.0.0
